requirements.txt
asgiref==3.11.1
certifi==2026.2.25
distlib==0.4.0
Django==6.0.3
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
djangorestframework==3.17.0
djangorestframework-simplejwt==5.5.1
filelock==3.25.2
packaging==26.0
pipenv==2026.1.0
platformdirs==4.9.4
PyJWT==2.12.1
PyMySQL==1.1.2
python-discovery==1.1.3
python-dotenv==1.2.2
setuptools==82.0.1
sqlparse==0.5.5
tzdata==2025.3
virtualenv==21.2.0
